Remove deprecated @usecase
This removes uses of @usecase. This is deprecated in Scala 2.13.x.
Additionally, scala generious spurious warnings if you use @usecase in
a non-abstract class. (It thinks that this is an abstract member of
something concrete.)

The guidance from upstream Scala is to be explicit about how to use
your API via @example.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>