Change when/switch thunk type to Any (#1308)
* Change when thunks return type to Any

Changes the type of the thunk for when and WhenContext methods from
call-by-name Unit to call-by-name Any. This prevents a
warning (-Ywarn-value-discard) where a when thunk is returning
something other than Unit that is then discarded, e.g., another
WhenContext.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Change switch thunk return to type to Any

Changes the type of switch thunks from call-by-name Unit to
call-by-name Any. This prevents a warning (-Ywarn-value-discard) when
the internals of a switch block return something other than Unit which
is then discarded.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>